2.2 voucher applied
-. user can finish transaction
-. update LoadAll Order for customer into 1 API + for merchant 1 API(status handle by frontend)

-. Voucher : loadAll, create, delete, apply('voucherCode'), cancel(empty body json)

======
-. exception & response voucher
-. /accOrRejectOrder ( /merchant => /order)
-. entity OTM MTO Voucher, orphanRemoval
-.

+++ src/main/java/com/demo/lookopediaSinarmas/controller/CartController.java
@@ -80,7 +80,7 @@ public class CartController {
 	}
 	
 	@GetMapping("/getTotalPrice/{user_id}")
-	public int getTotalPrice(@PathVariable Long user_id){
+	public double getTotalPrice(@PathVariable Long user_id){
 		return cartService.returnTotalPrice(user_id);
 	}
